Addressing review comments

+++ core/src/main/java/io/plaidapp/core/data/PlaidItem.kt
@@ -22,11 +22,10 @@ package io.plaidapp.core.data
 abstract class PlaidItem(
     @Transient open val id: Long,
     @Transient open val title: String? = null,
-// can't be final as some APIs use different serialized names
     @Transient open var url: String? = null
 ) {
     var dataSource: String? = null
     var page: Int = 0
-    var weight: Float = 0.toFloat() // used for sorting
+    var weight: Float = 0F // used for sorting
     var colspan: Int = 0
 }
